some changes in homework05

+++ Homework05/included/mean.hpp
@@ -22,5 +22,5 @@ public:
 
 private:
 	double m_mean;
-	int strLength;
+	size_t strLength;
 };
